# web/openapi/runtime.openapi.json
{
  "components": {
    "schemas": {
      "ActionRequest": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "$ref": "#/components/schemas/UserAction"
          },
          "channel": {
            "enum": [
              "web",
              "cli",
              "api",
              "mock"
            ],
            "title": "Channel",
            "type": "string"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "channel",
          "session_id",
          "action"
        ],
        "title": "ActionRequest",
        "type": "object"
      },
      "ActionResponseData": {
        "additionalProperties": false,
        "properties": {
          "action_outcome": {
            "$ref": "#/components/schemas/UserActionOutcome"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProjectedActionResult"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "action_outcome",
          "result"
        ],
        "title": "ActionResponseData",
        "type": "object"
      },
      "ActionResponseEnvelope": {
        "additionalProperties": false,
        "description": "ResponseEnvelope specialization for structured user actions.",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ActionResponseData"
          },
          "fallback_text": {
            "title": "Fallback Text",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "response_id": {
            "title": "Response Id",
            "type": "string"
          },
          "schema_version": {
            "default": "phase0.sdui.v1",
            "title": "Schema Version",
            "type": "string"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "running",
              "completed",
              "blocked",
              "waiting_user",
              "failed",
              "no_capability_found",
              "cancelled",
              "confirmation_invalidated"
            ],
            "title": "Status",
            "type": "string"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          },
          "trace_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Summary"
          },
          "ui": {
            "discriminator": {
              "mapping": {
                "binding_required_card": "#/components/schemas/UIComponent",
                "confirm_card": "#/components/schemas/ConfirmCard",
                "none": "#/components/schemas/UIComponent",
                "operator_handback_card": "#/components/schemas/UIComponent"
              },
              "propertyName": "component_type"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/ConfirmCard"
              },
              {
                "$ref": "#/components/schemas/UIComponent"
              }
            ],
            "title": "Ui"
          }
        },
        "required": [
          "response_id",
          "task_id",
          "session_id",
          "status",
          "message",
          "fallback_text",
          "ui",
          "data",
          "trace_id"
        ],
        "title": "ActionResponseEnvelope",
        "type": "object"
      },
      "CancelUserAction": {
        "additionalProperties": false,
        "properties": {
          "action_type": {
            "const": "cancel",
            "title": "Action Type",
            "type": "string"
          },
          "response_id": {
            "title": "Response Id",
            "type": "string"
          }
        },
        "required": [
          "action_type",
          "response_id"
        ],
        "title": "CancelUserAction",
        "type": "object"
      },
      "ConfirmCard": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "const": "confirm",
            "title": "Action",
            "type": "string"
          },
          "component_type": {
            "const": "confirm_card",
            "default": "confirm_card",
            "title": "Component Type",
            "type": "string"
          },
          "payload": {
            "$ref": "#/components/schemas/ConfirmCardPayload"
          },
          "reason_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason Code"
          },
          "target_system": {
            "anyOf": [
              {
                "enum": [
                  "oa",
                  "u8",
                  "hikvision_ivms",
                  "business_platform"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target System"
          }
        },
        "required": [
          "action",
          "payload"
        ],
        "title": "ConfirmCard",
        "type": "object"
      },
      "ConfirmCardPayload": {
        "additionalProperties": false,
        "description": "Runtime-owned payload contract for operation confirmation cards.",
        "properties": {
          "capability_id": {
            "title": "Capability Id",
            "type": "string"
          },
          "displayed_argument_values": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Displayed Argument Values",
            "type": "object"
          },
          "field_names": {
            "items": {
              "type": "string"
            },
            "title": "Field Names",
            "type": "array"
          },
          "operation_summary": {
            "title": "Operation Summary",
            "type": "string"
          },
          "target_system": {
            "anyOf": [
              {
                "enum": [
                  "oa",
                  "u8",
                  "hikvision_ivms",
                  "business_platform"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target System"
          }
        },
        "required": [
          "capability_id",
          "operation_summary",
          "target_system",
          "field_names",
          "displayed_argument_values"
        ],
        "title": "ConfirmCardPayload",
        "type": "object"
      },
      "ConfirmUserAction": {
        "additionalProperties": false,
        "properties": {
          "action_type": {
            "const": "confirm",
            "title": "Action Type",
            "type": "string"
          },
          "confirmed": {
            "const": true,
            "title": "Confirmed",
            "type": "boolean"
          },
          "response_id": {
            "title": "Response Id",
            "type": "string"
          }
        },
        "required": [
          "action_type",
          "response_id",
          "confirmed"
        ],
        "title": "ConfirmUserAction",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HandleRequest": {
        "additionalProperties": false,
        "properties": {
          "channel": {
            "enum": [
              "web",
              "cli",
              "api",
              "mock"
            ],
            "title": "Channel",
            "type": "string"
          },
          "client_capabilities": {
            "additionalProperties": true,
            "title": "Client Capabilities",
            "type": "object"
          },
          "client_request_id": {
            "anyOf": [
              {
                "maxLength": 96,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Request Id"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "channel",
          "session_id",
          "message"
        ],
        "title": "HandleRequest",
        "type": "object"
      },
      "ProjectedActionResult": {
        "additionalProperties": {},
        "description": "Dynamic business result after app.runtime.response_projection.project_response_data applies CapabilitySpec.output_schema; this OpenAPI shape is not an exposure allowlist.",
        "title": "ProjectedActionResult",
        "type": "object"
      },
      "RejectUserAction": {
        "additionalProperties": false,
        "properties": {
          "action_type": {
            "const": "reject",
            "title": "Action Type",
            "type": "string"
          },
          "response_id": {
            "title": "Response Id",
            "type": "string"
          }
        },
        "required": [
          "action_type",
          "response_id"
        ],
        "title": "RejectUserAction",
        "type": "object"
      },
      "ResponseEnvelope": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "fallback_text": {
            "title": "Fallback Text",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "response_id": {
            "title": "Response Id",
            "type": "string"
          },
          "schema_version": {
            "default": "phase0.sdui.v1",
            "title": "Schema Version",
            "type": "string"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "running",
              "completed",
              "blocked",
              "waiting_user",
              "failed",
              "no_capability_found",
              "cancelled",
              "confirmation_invalidated"
            ],
            "title": "Status",
            "type": "string"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          },
          "trace_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Summary"
          },
          "ui": {
            "discriminator": {
              "mapping": {
                "binding_required_card": "#/components/schemas/UIComponent",
                "confirm_card": "#/components/schemas/ConfirmCard",
                "none": "#/components/schemas/UIComponent",
                "operator_handback_card": "#/components/schemas/UIComponent"
              },
              "propertyName": "component_type"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/ConfirmCard"
              },
              {
                "$ref": "#/components/schemas/UIComponent"
              }
            ],
            "title": "Ui"
          }
        },
        "required": [
          "response_id",
          "task_id",
          "session_id",
          "status",
          "message",
          "fallback_text",
          "ui",
          "trace_id"
        ],
        "title": "ResponseEnvelope",
        "type": "object"
      },
      "UIComponent": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "anyOf": [
              {
                "enum": [
                  "confirm",
                  "bind_required",
                  "clarify_scope",
                  "none"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Action"
          },
          "component_type": {
            "enum": [
              "none",
              "operator_handback_card",
              "binding_required_card"
            ],
            "title": "Component Type",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "reason_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason Code"
          },
          "target_system": {
            "anyOf": [
              {
                "enum": [
                  "oa",
                  "u8",
                  "hikvision_ivms",
                  "business_platform"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target System"
          }
        },
        "required": [
          "component_type"
        ],
        "title": "UIComponent",
        "type": "object"
      },
      "UserAction": {
        "discriminator": {
          "mapping": {
            "cancel": "#/components/schemas/CancelUserAction",
            "confirm": "#/components/schemas/ConfirmUserAction",
            "reject": "#/components/schemas/RejectUserAction"
          },
          "propertyName": "action_type"
        },
        "oneOf": [
          {
            "$ref": "#/components/schemas/ConfirmUserAction"
          },
          {
            "$ref": "#/components/schemas/RejectUserAction"
          },
          {
            "$ref": "#/components/schemas/CancelUserAction"
          }
        ]
      },
      "UserActionOutcome": {
        "enum": [
          "accepted",
          "action_gate_unavailable",
          "no_pending_action",
          "action_binding_incomplete",
          "action_reference_mismatch",
          "action_pending_changed",
          "action_already_claimed",
          "action_stale",
          "action_version_conflict",
          "cancelled",
          "confirmation_invalidated"
        ],
        "type": "string"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "EternalAI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/runtime/action": {
      "post": {
        "operationId": "handle_action_api_v1_runtime_action_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponseEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Handle Action"
      }
    },
    "/api/v1/runtime/handle": {
      "post": {
        "operationId": "handle_api_v1_runtime_handle_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HandleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Handle"
      }
    }
  }
}